chore: fix popper click
