+++ nx.json
@@ -15,6 +15,9 @@
     "build": {
       "dependsOn": [
         "^build"
+      ],
+      "outputs": [
+        "{projectRoot}/dist"
       ]
     },
     "clean": {
