+++ packages/components/dismissable-layer/src/DismissableLayer.ts
@@ -4,15 +4,12 @@ import {
   useEscapeKeydown,
   useForwardRef,
 } from '@oku-ui/use-composable'
-import type { OkuElement, PrimitiveProps } from '@oku-ui/primitive'
 import { Primitive, primitiveProps } from '@oku-ui/primitive'
-import type { Ref } from 'vue'
 import {
   computed,
   defineComponent,
   h,
   mergeProps,
-  nextTick,
   onMounted,
   onUnmounted,
   reactive,
@@ -27,107 +24,10 @@ import {
   useFocusoutSide,
   usePointerdownOutside,
 } from './util'
-
-export const dismissableLayerContext = reactive({
-  layersRoot: new Set<DismissableLayerElement>(),
-  layersWithOutsidePointerEventsDisabled: new Set<DismissableLayerElement>(),
-  branches: new Set<DismissableLayerElement>(),
-})
-
-export const INJECT_UPDATE = 'dismissableLayer.update'
-export const POINTER_DOWN_OUTSIDE = 'dismissableLayer.pointerDownOutside'
-export const FOCUS_OUTSIDE = 'dismissableLayer.focusOutside'
+import type { DismissableLayerNativeElement, FocusBlurCaptureEvent, FocusCaptureEvent, PointerdownCaptureEvent } from './props'
+import { DISMISSABLE_NAME, INJECT_UPDATE, dismissableLayerContext, dismissableLayerProps } from './props'
 
 let originalBodyPointerEvents: string
-
-export const DISMISSABLE_NAME = 'OkuDismissableLayer'
-export const DismissableLayerProvideKey = Symbol('DismissableLayerProvide')
-
-export type DismissableLayerNativeElement = OkuElement<'div'>
-export type DismissableLayerElement = HTMLDivElement
-
-export type DismissableLayerProvideValue = {
-  layers: Ref<Set<DismissableLayerElement>>
-  layersWithOutsidePointerEventsDisabled: Ref<Set<DismissableLayerElement>>
-  branches: Ref<Set<DismissableLayerElement>>
-}
-
-export type PointerdownOutsideEvent = CustomEvent<{
-  originalEvent: PointerEvent
-}>
-export type FocusoutSideEvent = CustomEvent<{ originalEvent: FocusEvent }>
-export type FocusCaptureEvent = CustomEvent<{ originalEvent: FocusEvent }>
-export type FocusBlurCaptureEvent = CustomEvent<{ originalEvent: FocusEvent }>
-export type PointerdownCaptureEvent = CustomEvent<{
-  originalEvent: PointerEvent
-}>
-
-export interface DismissableLayerProps extends PrimitiveProps {
-  /**
-   * When `true`, hover/focus/click interactions will be disabled on elements outside
-   * the `DismissableLayer`. Users will need to click twice on outside elements to
-   * interact with them: once to close the `DismissableLayer`, and again to trigger the element.
-   */
-  disableOutsidePointerEvents?: boolean
-}
-
-export type DismissableLayerEmits = {
-  /**
-  * Event handler called when the escape key is down.
-  * Can be prevented.
-  */
-  escapeKeyDown: [event: KeyboardEvent]
-  /**
-  * Event handler called when the a `pointerdown` event happens outside of the `DismissableLayer`.
-  * Can be prevented.
-  */
-  pointerdownOutside: [event: PointerdownOutsideEvent]
-  /**
-   * Event handler called when the focus moves outside of the `DismissableLayer`.
-   * Can be prevented.
-   */
-  focusoutSide: [event: FocusoutSideEvent]
-  /**
-  * Event handler called when an interaction happens outside the `DismissableLayer`.
-  * Specifically, when a `pointerdown` event happens outside or focus moves outside of it.
-  * Can be prevented.
-  */
-  interactOutside: [event: PointerdownOutsideEvent | FocusoutSideEvent]
-  /**
-  * Handler called when the `DismissableLayer` should be dismissed
-  */
-  dismiss: []
-  focusCapture: [event: FocusCaptureEvent]
-  blurCapture: [event: FocusBlurCaptureEvent]
-  pointerdownCapture: [event: PointerdownCaptureEvent]
-}
-
-export const dismissableLayerProps = {
-  props: {
-    disableOutsidePointerEvents: {
-      type: Boolean,
-      default: false,
-    },
-  },
-
-  emits: {
-    // eslint-disable-next-line unused-imports/no-unused-vars
-    escapeKeyDown: (event: KeyboardEvent) => true,
-    // eslint-disable-next-line unused-imports/no-unused-vars
-    pointerdownOutside: (event: PointerdownOutsideEvent) => true,
-    // eslint-disable-next-line unused-imports/no-unused-vars
-    focusoutSide: (event: FocusoutSideEvent) => true,
-    // eslint-disable-next-line unused-imports/no-unused-vars
-    interactOutside: (event: PointerdownOutsideEvent | FocusoutSideEvent) => true,
-    dismiss: () => true,
-    // eslint-disable-next-line unused-imports/no-unused-vars
-    focusCapture: (event: FocusCaptureEvent) => true,
-    // eslint-disable-next-line unused-imports/no-unused-vars
-    blurCapture: (event: FocusBlurCaptureEvent) => true,
-    // eslint-disable-next-line unused-imports/no-unused-vars
-    pointerdownCapture: (event: PointerdownCaptureEvent) => true,
-  },
-}
 
 const DismissableLayer = defineComponent({
   name: DISMISSABLE_NAME,
@@ -188,7 +88,7 @@ const DismissableLayer = defineComponent({
 
       if (!event.defaultPrevented)
         emit('dismiss')
-    }, ownerDocument.value)
+    }, ownerDocument)
 
     const focusoutSide = useFocusoutSide((event) => {
       const target = event.target as HTMLElement
@@ -204,7 +104,7 @@ const DismissableLayer = defineComponent({
 
       if (!event.defaultPrevented)
         emit('dismiss')
-    }, ownerDocument.value)
+    }, ownerDocument)
 
     useEscapeKeydown((event) => {
       const isHighestLayer = index.value === dismissableLayerContext.layersRoot.size - 1
@@ -214,9 +114,9 @@ const DismissableLayer = defineComponent({
 
       emit('escapeKeyDown', event)
 
-      if (!event.defaultPrevented && props.onDismiss)
+      if (!event.defaultPrevented)
         emit('dismiss')
-    }, ownerDocument.value)
+    }, ownerDocument)
 
     watchEffect(async (onInvalidate) => {
       if (!node.value)
@@ -228,7 +128,6 @@ const DismissableLayer = defineComponent({
             = ownerDocument.value.body.style.pointerEvents
           ownerDocument.value.body.style.pointerEvents = 'none'
         }
-        await nextTick()
         dismissableLayerContext.layersWithOutsidePointerEventsDisabled.add(node.value)
       }
