+++ packages/components/dismissable-layer/src/index.ts
@@ -2,10 +2,13 @@ export * from './DismissableLayerBranch'
 
 export {
   OkuDismissableLayer,
-  dismissableLayerProps,
 } from './DismissableLayer'
 
 export type {
   DismissableLayerProps,
   DismissableLayerEmits,
-} from './DismissableLayer'
+} from './props'
+
+export {
+  dismissableLayerProps,
+} from './props'
