+++ packages/core/use-composable/src/useEscapeKeydown.ts
@@ -1,19 +1,20 @@
-import { watchEffect } from 'vue'
+import type { Ref } from 'vue'
+import { ref, watchEffect } from 'vue'
 
 /**
  * Listens for when the escape key is down
  */
 function useEscapeKeydown(
   onEscapeKeyDownProp?: (event: KeyboardEvent) => void,
-  ownerDocument: Document = globalThis?.document,
+  ownerDocument: Ref<Document> = ref(globalThis?.document),
 ) {
   watchEffect((onInvalidate) => {
     const handleKeyDown = (event: KeyboardEvent) => {
       if (event.key === 'Escape')
         onEscapeKeyDownProp?.(event)
     }
-    ownerDocument.addEventListener('keydown', handleKeyDown)
-    onInvalidate(() => ownerDocument.removeEventListener('keydown', handleKeyDown))
+    ownerDocument.value.addEventListener('keydown', handleKeyDown)
+    onInvalidate(() => ownerDocument.value.removeEventListener('keydown', handleKeyDown))
   })
 }
